Task 48 : Convert totalInTime value from Integer to Double

+++ src/main/java/com/springshortpath/app/payload/response/PathShortestTimeResponse.java
@@ -9,7 +9,7 @@ public class PathShortestTimeResponse {
 
     private String departureCity;
     private String arrivalCity;
-    private Integer totalInTime;
+    private Double totalInTime;
 
     public PathShortestTimeResponse(PathShortestTimeResponse pathShortestTimeResponse) {
         this.departureCity = pathShortestTimeResponse.getDepartureCity();
